+++ package.json
@@ -15,6 +15,7 @@
     "eslint": "7.18.0",
     "eslint-config-htmlacademy": "2.0.0"
   },
+  "type": "module",
   "license": "MIT",
   "dependencies": {},
   "scripts": {
